+++ src/com/company/homeworks/AreaOfRectangle/Shape.java
@@ -1,7 +1,7 @@
 package com.company.homeworks.AreaOfRectangle;
 
 public abstract class Shape {
-    public abstract int rectangleArea(double length, double breadth);
-    public abstract int squareArea(double side);
-    public abstract int circleArea(double radius);
+    public abstract double rectangleArea(double length, double breadth);
+    public abstract double squareArea(double side);
+    public abstract double circleArea(double radius);
 }
